Autocomplete more complete - Condition UC Much improved
The autocomplete button works better
Button should use ButtonDisabledScaling for text backdown.
TextBox - properly names the aux button.
theme - the setting of disabled scaling is strings.. not sure why it was done but was not working. need to check with EDD

ConditionfilterUC now takes a variable structure to give help and default condition on each variable.  Now autoselects the default condition.