Scroll bar improvement in performance
Isolated that the Visible=true was causing the slowdown, in CalculateThumb. Calling it even if it was already on was killing the performance.
So back out the  suspend from the PanelDataGridViewScroll and only set Visible when it needs to change.
This makes it perform to normal speed, even if you add one row at a time.