Nerf the message box so it allows more space for the scroll bar
Basic problem with richtextbox that its estimates are circular - its using the current text situation (based on text box size) to determine if to turn on/off the scroll bar - which takes space. Really should set the text box to max, determine if it fits, then turn the scroll bar on.